Normalize NI USB timeouts for LabRAD
LabRAD timeout values can arrive as floats in milliseconds. Convert timeout values to integers before passing them to PyUSB/libusb; reproduced via GPIB Device Manager startup against the running GPIB Bus server.

+++ niusbhs/__init__.py
@@ -566,7 +566,7 @@ def _env_flag(name: str) -> bool:
 
 
 def _transfer_timeout_ms(timeout_ms: int) -> int:
-    return timeout_msecs(timeout_ms * 1000) or 0
+    return int(timeout_msecs(int(timeout_ms) * 1000) or 0)
 
 
 __all__ = [
